requirements.txt
pillow==11.3.0
wheel==0.45.1